change db id type to bigint

+++ DDL.sql
@@ -1,6 +1,6 @@
 CREATE TABLE my_db.cities
 (
-    id          int          NOT NULL AUTO_INCREMENT,
+    id          bigint          NOT NULL AUTO_INCREMENT,
     name        VARCHAR(30) UNIQUE,
     description VARCHAR(100) NOT NULL,
     PRIMARY KEY (id)
